+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
 
-## 1.0.2 - 2019-08-23
+## 1.0.3 - 2019-08-26
 ### Added
   - Sorting of admin panel table.
+  - Fetching backup files now support Hidex files.
 
 ### Changed
   - Resolve minor spelling mistake in tooltip for delete and edit buttons on admin panel.